ci: Ignore RUSTSEC-2020-0071 in deny.toml
The advisory RUSTSEC-2020-0071 (alias
RUSTSEC-2020-26235) does not apply to the
`time` dependency in `chrono`, which is a
dependency of `osmosis-std`. After some
investigation it seems the advisory can
safely be ignored, but care must be taken
that it is not ignored for any other
crates, should it appear anywhere else.

Signed-off-by: Bobby <bobby@apollo.farm>